the data isn't coming correct on the dashboard bhr the assigned branches, branhces covered this month, total visits this month, and yeah remove approved reports, completion rate this month
these should be fetched live too:

assigned branches should simply show the number of branches assigned to the bhr,

then the branches. covered this month should show the totla number of branches. for whom atleast once a form is submitted in the particular month

in total visits should simply show the total number of forms submitted that month

and completion rate is simply the (branches coveredthat month / branches assigned )